requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.3
streamlit==1.32.0
pandas==2.2.0
plotly==5.19.0
psycopg2-binary==2.9.9
